% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bpp_mcmc.R
\name{as.bpp}
\alias{as.bpp}
\title{Importing BPP mcmc trace.}
\usage{
as.bpp(trees, log)
}
\arguments{
\item{trees}{The name of "trees" file or an object of class \code{multiPhylo}.}

\item{log}{The name of "log" file or a data frame.}
}
\description{
Imports mcmc trace of BPP from files where its components are stored.
}
