% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxtree.R
\name{boxtree}
\alias{boxtree}
\title{Box depiction of species tree.}
\usage{
boxtree(
  tree,
  gap = 0.1,
  col = "grey",
  border = "black",
  lwd = 3,
  cex = 1.5,
  round = 0.75,
  show.tip.label = TRUE,
  tip.label = NULL,
  tip.cex = cex,
  tip.font = 3,
  tip.offset = 1.1,
  root.lwd = 1,
  timescale = TRUE,
  axis.lwd = lwd/2,
  labels = NULL,
  digits = 4,
  lab.col = "white",
  lab.border = "black",
  lab.lwd = lwd/2,
  lab.cex = 1,
  lab.font = 1,
  xlab = "Tau",
  ylab = "Theta",
  xlim,
  ylim,
  mai,
  direction,
  device,
  file,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{tree}{Tree imported by \code{read_msc_tree}, i.e., \code{phylo} object with the additional component \code{params},
or name of a file, which can be imported by \code{read_msc_tree}.}

\item{gap}{The minimum gap between branches, specified as a proportion of the maximum branch width,
but then automatically adjusted.}

\item{col}{Color of branch background, can be a vector ordered as branches in \code{tree$edge}.}

\item{border}{Color of branch outline.}

\item{lwd}{Width of branch outline.}

\item{cex}{Size of axis labels, serves also for derivation of size of axis annotation.}

\item{round}{How much rounded are branch corners. Specified as a proportion of the minimum of branch
lengths and widths, which is then used as a diameter of the circle defining the curvature.}

\item{show.tip.label}{Logical, whether to show tip labels}

\item{tip.label}{Vector of tip labels, used if the desired tip labels differ from \code{tree$tip.label},
the vector is named by \code{tree$tip.label}, however, to establish mutual correspondence of the labels.}

\item{tip.cex}{Size of tip labels.}

\item{tip.font}{Font of tip labels.}

\item{tip.offset}{A multiplayer of space between tree tips and species labels.}

\item{root.lwd}{A multiplayer of the width of the root line.}

\item{timescale}{Logical, whether to display time scale axis.}

\item{axis.lwd}{Width of the axis line.}

\item{labels}{Character, what is displayed at branch labels, options are \code{"theta"}, \code{"tau"} or \code{"gdi"}.}

\item{digits}{Numeric, the number of decimal places for parameters to be displayed at node labels.}

\item{lab.col}{Color of branch label background.}

\item{lab.border}{Color of branch label outline.}

\item{lab.lwd}{Width of branch label outline}

\item{lab.cex}{Size of branch label}

\item{lab.font}{Font of branch label}

\item{xlab, ylab}{Labels of x and y axes}

\item{xlim, ylim}{Limits of x and y axes}

\item{mai}{Size of outer margins in inches, recycled if necessary}

\item{direction}{Tree direction (only 'rightwards' is implemented now)}

\item{device}{Character, type of the graphical device. If not specified (default), either \code{"quartz"} or \code{"x11"}
is chosen (depending on the platform). If \code{NULL}, the objects are plotted into the current device. If it is
\code{"pdf"}, the pdf file is created, whose name is specified by the \code{file} argument.}

\item{file}{Character, the name of pdf file.}

\item{width}{Width of graphical device.}

\item{height}{Height of graphical device.}
}
\description{
Plots species tree with branches represented by rectangles (boxes) whose sides are proportional to multispecies coalescent parameters ('tau' and 'theta'), expressed in comparable units (expected number of mutations, see Details).
}
\details{
Both divergence times ('taus') and population size parameters ('thetas') are
in units of the expected number of substitutions per site. However, taus correspond to
tip-node distances, whereas thetas to tip-node-tip distances, so the thetas are divided
by two to be in scale in the figure.
}
