% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_clades.R
\name{collapse_clades}
\alias{collapse_clades}
\title{Clades collapsed to branches.}
\usage{
collapse_clades(tree, clades)
}
\arguments{
\item{tree}{An object of class \code{phylo} or \code{multiPhylo} or a character string with tree in newick format
or a name of file containing the tree(s).}

\item{clades}{A data frame listing tip labels (1st column) and clade labels (2nd column) or a name of mapping file.}
}
\value{
An object of class \code{phylo} of \code{multiPhylo} specifying a tree with a single tip per clade
and terminal branch lengths equal to the mean root-to-tip distances in the clades.
}
\description{
Collapses specified clades to single branches.
}
\details{
The specified clades must be monophyletic in the tree(s), but may cover only a part of the tree(s).
}
