% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bpp_mcmc.R
\name{combine_bpp_runs}
\alias{combine_bpp_runs}
\title{Combine BPP runs.}
\usage{
combine_bpp_runs(runs)
}
\arguments{
\item{runs}{The list of lists returned by \code{read_bpp_mcmc}.}
}
\value{
The list with the same components as the input objects,
containing pooled posterior samples.
}
\description{
Combine MCMC traces from independent replicates of the same analysis in BPP.
}
