% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate.R
\name{concatenate}
\alias{concatenate}
\title{Sequence concatenation.}
\usage{
concatenate(
  seq,
  nloci = -1L,
  class = "matrix",
  toupper = TRUE,
  missing = "-",
  allele
)
}
\arguments{
\item{seq}{The list of alignments with row names indicating individuals
or a single such alignment (possibly concatenated multi-locus data).}

\item{nloci}{The number of loci to be imported (counted from the beginning). The default means "all loci".}

\item{class}{The class of output object, either 'matrix', 'DNAbin' or 'DNAStringSet'.}

\item{toupper}{Logical, whether to represent bases by upper case letters. Always \code{TRUE} if \code{allele} is specified.}

\item{missing}{Character, how to represent missing data.}

\item{allele}{The allele identifier distinguishing haplotypes of the same individual.
If supplied, the individual diploid genotypes are collapsed into a single sequence.}
}
\value{
A single concatenated alignment with attributes \code{part} (a two-column matrix, which defines
paritioning into loci) and possibly \code{bedtable} and \code{snps} (if these were included in the input data).
}
\description{
Concatenates multi-locus sequences into a single alignement, while preserving their atributes.
}
