% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPD.R
\name{CPD}
\alias{CPD}
\title{The central posterior density interval.}
\usage{
CPD(x, p = 0.95)
}
\arguments{
\item{x}{A numeric vector representing posterior sample of a model parameter.}

\item{p}{The proportion of posterior density be included in the interval, default is 0.95.
It implies the maignitude of exterme quantiles of posterior density, \code{q = (1 - p) / 2}.}
}
\value{
A numeric vector with lower and upper limit of CPD interval.
}
\description{
Estimates the central posterior density interval.
}
\details{
The central posterior density interval corresponds to the range of values after trimming
of tails corresponding to the extreme quantiles of posterior density.
}
