% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgamma.R
\name{dinvgamma}
\alias{dinvgamma}
\title{Inverse gamma distribution.}
\usage{
dinvgamma(x, shape, scale, log = FALSE)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{shape}{Shape parameter, must be greater than one.}

\item{scale}{Scale parameter, must be positive.}

\item{log, log.p}{Logical, whether logarithms of densities/probabilities are returned.}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}

\item{lower.tail}{Logical, if TRUE (default), probabilities are ---, otherwise, ---.}
}
\description{
Density function, distribution function, quantile function and random generation
for the Inverse gamma distribution with parameters \code{shape} and \code{scale}.
}
\details{
The \code{qinvgamma} function is currently a copy of implementation from \code{invgamma} package.
}
