% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_marglik.R
\name{estimate_marglik}
\alias{estimate_marglik}
\title{Marginal likelihood by thermodynamic integration.}
\usage{
estimate_marglik(files, weights, pattern = "\\\\.out")
}
\arguments{
\item{files}{names of the .out files or a folder containing them.}

\item{weights}{a numeric vector with Gauss-Legendre weights (in the increase order of their betas)
or a name of .csv file containing them (generated by BPP's bfdriver)
or a data frame with the same columns as the content of the .csv file.}

\item{pattern}{regular expression allowing to recognize ".out" files, if \code{files} is a folder name.}
}
\value{
The logarithm of the marginal likelihood.
}
\description{
Extracts the estimate of marginal likelihood of a model from the outputs of
a series of MCMC runs conducted in the thermodynamic integration procedure.
}
