% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_priors.R
\name{estimate_otus}
\alias{estimate_otus}
\title{Operational taxonomic units.}
\usage{
estimate_otus(seq, seqmap = NULL, allele = NULL, model = "JC69")
}
\arguments{
\item{seq}{The list of locus-specific alignments or a single such alignment (possibly
with concatenated multi-locus data). The row names must be interpretable as individual labels
(possibly with helpof \code{allele} or \verb{seqmap arguments}).}

\item{seqmap}{A data frame mapping names of sequences (1st column) to those of individuals (2nd column).}

\item{allele}{An allele identifier, regular expression distinguishing sequences from the same individual.}

\item{model}{The nucleotide substitution model indicated as in \code{\link[ape:dist.dna]{ape::dist.dna()}}.}
}
\value{
A data frame with columns \code{"Individuals"} and \code{"Species"}.
}
\description{
Provides delimitation of operational taxonomic units (OTUs, provisional species) using
the maximum cross-section branch length criterion on ad hoc created average linkage (UPGMA) dendrogram.
}
\details{
The function provides provisional delimitation of operational taxonomic units
intended to be used as a rough guide to inform pre-estimation of prior parameters.
It uses a simple "maximum cross-section length" method, which bulids ultrametric UPGMA dendrogram
from the matrix of pairwise inter-individual genetic distances and then cuts it into clusters
at some distance from the tips. The distance is chosen to maximize the average length of branches
that are cut by the cross-section line.
}
