% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_marglik.R
\name{estimate_sdratio}
\alias{estimate_sdratio}
\title{Bayes factor estimated by Savage-Dickey ratio.}
\usage{
estimate_sdratio(prior, pars, posterior, null, log = FALSE)
}
\arguments{
\item{prior}{The type of prior distribution, must be a character string
for which quantile function named \code{paste0("p", prior)} is available.}

\item{pars}{A numeric vector with parameters of the prior distribution.
They are used as arguments of the quantile function.}

\item{posterior}{A posterior sample of the parameter in question.}

\item{null}{A numeric vector giving the null region, i.e. the interval close to the null value.
If just a single number is supplied, the other limit is supposed to be zero.}

\item{log}{Logical, whether to return logarithm of the value.}
}
\value{
The value of Savage-Dickey ratio.
}
\description{
Calculates Savage-Dickey ratio as an approximation of Bayes factor comparing the posterior sample
of a parameter to its prior probability density in a nested null model.
}
