% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_priors.R
\name{estimate_starting_tree}
\alias{estimate_starting_tree}
\title{Starting tree for species tree estimation.}
\usage{
estimate_starting_tree(
  seq,
  method = c("nj", "upgma"),
  imap = NULL,
  seqmap = NULL,
  allele = NULL,
  model = "JC69"
)
}
\arguments{
\item{seq}{The list of locus-specific alignments or a single such alignment (possibly
with concatenated multi-locus data). The row names must be interpretable as individual labels
(possibly with helpof \code{allele} or \verb{seqmap arguments}). Alternatively, it can be a matrix of
average genetic distances between (candidate) species, supplied as an object of class \code{dist}
or a symmetric matrix of mode \code{numeric}.}

\item{method}{The tree estimation method, either \code{"nj"} (neighbor joining) or \code{"upgma"}
(unweighted pair group method with arithmetic mean); corresponds to options \code{"single"}
and \code{"average"}, respectively, in \code{\link[stats:hclust]{stats::hclust()}}.}

\item{imap}{A matrix or data frame (1st column individuals, 2nd column species) or a name of imap file.}

\item{seqmap}{A data frame mapping names of sequences (1st column) to those of individuals (2nd column).}

\item{allele}{An allele identifier, regular expression distinguishing sequences from the same individual.}

\item{model}{The nucleotide substitution model indicated as in \code{\link[ape:dist.dna]{ape::dist.dna()}}.}
}
\value{
The starting tree topology in Newick format.
}
\description{
Estimates starting tree for species tree estimation (analyses A11 or A01)
from average genetic distances between (candidate) species.
}
