% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_statistic.R
\name{estimate_statistic}
\alias{estimate_statistic}
\title{Additional statistics.}
\usage{
estimate_statistic(bpp, which, args = NULL)
}
\arguments{
\item{bpp}{An object of class \code{bpp} as created by \link{read_bpp_mcmc} or \link{as.bpp},
or a data frame with postrior sample of parameters.}

\item{which}{Character, which statistic to estimate.
Currently limited to the names of functions defined in \code{bpptools}.}

\item{args}{A named list of arguments to the estimator function.}
}
\description{
Estimates additional statistics from sampled parameter values.
}
