% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_msc_tree.R
\name{export_msc_tree}
\alias{export_msc_tree}
\title{Multispecies coalescent tree.}
\usage{
export_msc_tree(
  target,
  trees,
  params,
  file,
  treesfile,
  fun = "mean",
  interval = "HPD",
  p = 0.95,
  gdi = FALSE,
  figtree = FALSE,
  constraint
)
}
\arguments{
\item{target}{The tree topology in Newick format obtained by \code{read_bpp_tree}
or a name of BPP control (.ctl) or output (.out) file.}

\item{trees}{A posterior sample of rooted trees in a \code{multiPhylo} object or a list of \code{phylo} objects
or a name of file with the posterior sample.}

\item{params}{A data frame with posterior sample of model parameters.}

\item{file}{A name of the file with the final MSC-annotated tree.}

\item{treesfile}{A file name for export of the annotated posterior sample of trees.}

\item{fun}{The summary function, \code{"mean"} (default) or \code{"median"}.}

\item{interval}{Either \code{"HPD"} (default) or \code{"CPD"}.}

\item{p}{The proportion of posterior density be included in the interval, default is 0.95.}

\item{figtree}{Logical, whether to export the final tree and the posterior sample.
as .nexus files readable by FigTree.}

\item{constraint}{a character vector specifying bottom-up succession of tips when plotting the tree}
}
\value{
\code{read_msc_tree} returns an object of class \code{phylo} with an additional component \code{params}, which is
a data frame including  MSC annotations: divergence times (tau) and mutation-scaled population sizes (theta).
Every row corresponds to a single node and the parameters relate to the branch leading to the node.
The root species has its theta but not tau.
}
\description{
Exports species tree annotated by the multispecies coalescent parameters in nexus metacomment format.
}
