% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_spdelim.R
\name{find_spdelim}
\alias{find_spdelim}
\title{Species delimitation.}
\usage{
find_spdelim(
  trees,
  file,
  type = c("M50", "MAP", "MC"),
  sep = "-",
  collapse = 0
)
}
\arguments{
\item{trees}{A posterior sample of rooted trees in a \code{multiPhylo} object or a list of \code{phylo} objects
or a name of file with the posterior sample.}

\item{file}{A name of the output file.}

\item{type}{The type of delimitation, either \code{"M50"} (majority consensus, default),
\code{"MAP"} (maximum aposteriori) or \code{"MC"} (maximum credibility), see Details for explanantion.}

\item{sep}{Symbol used to concatenate candidate species names.}

\item{collapse}{The divergence time (in appropriate units) approximating zero. In analyses based
on the rjMCMC algorithm (used by BPP) it is zero. In analyses based on birth-death-collapse prior
(in BEAST2) it is a small value identical to "collapse height" set in the analysis.}
}
\value{
A data frame with names of species in the chosen delimitation and their posterior probabilities
and a tab-delimited file with this output (if \code{file} is specified).
}
\description{
Finds species delimitation representative for a posterior sample.

\code{find_spdelim} a wrapper function joining the following functions into a single pipeline.

\code{speciesDA} takes a posterior sample of trees and extracts a sample of species delimitations out of it.
It returns a numeric matrix with rows for the sampled delimitation and columns for the tree tips.
The tips with the same number in given row are conspecific in that sampled delimitation.

\code{speciesPP} takes an output of \code{speciesDA} and converts it to a list with components
\code{"delims"} (data frame with delimitations and their PPs), \code{"species"} (data frame with species and PPs)
and \code{"tips"} (candidate species ordered as in column names of \code{speciesDA} output).

\code{speciesM50}, \code{speciesMAP}, \code{speciesMC} functions choosing majority consensus, the maximum a posteriori
and the maximum credibility species delimtation, respectively. They all return a data frame with species
in the 1st column and their PPs in the 2nd column and with an attribute \code{"pp"}, which is PP
of the delimitation as a whole
}
\details{
The majority consensus delimitation adds the most supported species unless they overlap
(they are in conflict) with any of the species that were already included. The maximum a posteriori
delimitation pick the set of species which was most frequently sampled as a whole. The maximum credibility
delimitation scores every sampled delimitation by the product of PPs of its constituent species.
}
