% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sptree.R
\name{find_sptree}
\alias{find_sptree}
\title{Species tree.}
\usage{
find_sptree(
  trees,
  type = c("MAP", "MCC", "CON"),
  file,
  thinning = 1,
  digits = 2
)
}
\arguments{
\item{trees}{A posterior sample of rooted trees in a \code{multiPhylo} object or a list of \code{phylo} objects
or an object of class \code{bpp} or a name of file with the posterior sample.}

\item{type}{The type of summary tree, either \code{"MAP"} (maximum aposteriori, default),
\code{"MCC"} (maximum clade credibility) or \code{"CON"} (majority consensus), see Details for explanantion.}

\item{file}{A name of the output file.}

\item{thinning}{A proportion of mcmc steps to be retained or a subsampling frequency (if \code{thin > 1}).}

\item{digits}{The number of digits in PP values.}
}
\value{
A \code{phylo} object with tree whose node labels indicate posterior probabilities of the clades.
}
\description{
Finds species tree representative for a posterior sample.

\code{find_sptree} a wrapper function joining the following functions into a single pipeline.

\code{treePP} takes a posterior sample of trees and converts it into a list with components
\code{"trees"} (data frame with species trees, their posterior probabilities and clade credibilities)
and \code{"clades"} (data frame with clades and their posterior probabilities).

\code{treeMAP}, \code{treeMC}, \code{treeCON} functions choosing the maximum a posteriori
the maximum credibility or the majority consensus species tree, respectively. They all return a \code{phylo} object
with tree whose node labels indicate PPs of the clades.
}
\details{
The maximum a posteriori tree is the most frequently sampled tree.
The maximum clade credibility tree scores every sampled tree by the product of PPs of its constituent clades.
The majority consensus tree adds the most supported clades unless they overlap (they are in conflict)
with any of the clades that were already included.
}
