% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdi.R
\name{gdi}
\alias{gdi}
\title{Genealogical divergence index.}
\usage{
gdi(params, tree, interval = "HPD", p = 0.95)
}
\arguments{
\item{params}{A data frame with multispecies coalescent parameters.}

\item{tree}{A tree in Newick format obtained by \code{read_bpp_tree}
or the name of BPP control (.ctl) or output (.out) file.}

\item{interval}{Either \code{"HPD"} (default) or \code{"CPD"}.}

\item{p}{The proportion of posterior density be included in the interval, default is 0.95.}
}
\value{
A list with components \code{"means"} (vector of posterior means), \code{"distributions"} (matrix
with posterior distributions) and \code{"interval"} (matrix with posterior density interval limits).
}
\description{
Calculates the genealogical divergence index.
}
\details{
Currently, it calculates gdi for every branch in the tree (internal or terminal)
under assumption of no postdivergence gene flow. The calculation for terminal branches
under models with gene flow is not implemented yet.
}
