% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collapsed_tree.R
\name{get_collapsed_tree}
\alias{get_collapsed_tree}
\title{Clades collapsed to branches.}
\usage{
get_collapsed_tree(tree, clades)
}
\arguments{
\item{tree}{An object of class \code{phylo} or a character string with tree in Newick format.}

\item{clades}{A data frame listing tip labels (1st column) and clade labels (2nd column) or a name of 'imap' file.}
}
\value{
An object of class \code{phylo} specifying a tree with a single tip per clade and the terminal branch length
equal to the mean distance between MRCA of the population and its tips in the original tree.
}
\description{
Collapses specified clades to single branches.
}
\details{
The species are required to be monophyletic in the tree, but they need not to cover the whole tree.
}
