% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bpp_ctl.R
\name{get_maxsamplesize}
\alias{get_maxsamplesize}
\title{Maximum sample sizes.}
\usage{
get_maxsamplesize(seq, imap, allele = NULL)
}
\arguments{
\item{seq}{The sequence file name or a list of alignments.}

\item{imap}{The imap file name or a data frame with the content of imap file.}

\item{allele}{An allele identifier, regular expression distinguishing sequences from the same individual.}
}
\value{
A named numeric vector with maximum sample sizes (the names are species labels).
}
\description{
Finds the maximum sample size (no. of sequences) per locus for each of the species.
}
