% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bpp_ctl.R
\name{get_nloci}
\alias{get_nloci}
\title{Maximum number of loci.}
\usage{
get_nloci(seqfile)
}
\arguments{
\item{seqfile}{The sequence file name.}
}
\value{
The number of loci.
}
\description{
Finds the maximum number of loci.
}
