% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bpp_ctl.R
\name{get_starting_tree}
\alias{get_starting_tree}
\title{Random starting tree.}
\usage{
get_starting_tree(species, imap)
}
\arguments{
\item{species}{A character vector of species names.}

\item{imap}{A data frame with a component \code{"Species"} or just two columns (2nd assumed to contain
species labels) or a name of 'imap' file.}
}
\value{
Starting tree topology in Newick format.
}
\description{
Creates a starting tree for species tree estimation (A11 or A01 analysis), using \code{\link[ape:rtree]{ape::rtree()}}.
}
