% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bpp_ctl.R
\name{get_topology}
\alias{get_topology}
\title{Tree topology.}
\usage{
get_topology(tree)
}
\arguments{
\item{tree}{An object of class \code{phylo} or a character string with tree in Newick format.}
}
\value{
A character string with tree topology (no branch lengths) in Newick format.
}
\description{
Writes tree topology in Newick format.
}
