% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPD.R
\name{HPD}
\alias{HPD}
\title{The highest posterior density interval.}
\usage{
HPD(x, p = 0.95)
}
\arguments{
\item{x}{A numeric vector, the posterior sample of a model paramater.}

\item{p}{The proportion of posterior density be included in the interval, default is 0.95.}
}
\value{
A numeric vector with lower and upper limit of HPD interval.
}
\description{
Estimates the highest posterior density interval.
}
\details{
"The 95\% HPD stands for highest posterior density interval and represents the most compact interval
on the selected parameter that contains 95\% of the posterior probability." (Drummond & Bouckaert 2015, p. 89)
}
