% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_seq.R
\name{import_fasta}
\alias{import_fasta}
\title{Importing fasta files.}
\usage{
import_fasta(
  file,
  nloci = -1L,
  class = "matrix",
  toupper = TRUE,
  part = NULL,
  rm.empty = FALSE,
  nlines,
  indiv,
  interleaved
)
}
\arguments{
\item{file}{Character, the name(s) of file(s) to be imported (each file corresponds to a single locus).}

\item{nloci}{The number of loci to be imported (the default \code{-1L} means all of them).}

\item{class}{Character, the output class, either \code{"matrix"} or \code{"DNAbin"}.}

\item{toupper}{Logical, whether to represent nucleotides by upper case letters.}

\item{part}{It can be a two column matrix indicating starts and ends of loci in the concatenated aligments
or a name of partition file formatted according to RAxML or MrBayes convention.}

\item{rm.empty}{Logical, whether to remove empty (e.g. all "-" or all "N") sequences.}

\item{nlines}{Not relevant, just for compatibility with \link{import_seq}.}

\item{indiv}{Not relevant, just for compatibility with \link{import_seq}.}

\item{interleaved}{Not relevant, just for compatibility with \link{import_seq}.}
}
\value{
A list of locus-specific alignments.
}
\description{
Importing sequence data from files in .fasta format.
}
