% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_imap.R
\name{import_imap}
\alias{import_imap}
\title{Importing classification information.}
\usage{
import_imap(file, header = TRUE, col.names, sep = "\\t")
}
\arguments{
\item{file}{A name of file in table format. It must contain either the columns labelled \code{"Individuals"}
and \code{"Species"} or these two and an additional column \code{"Sequence"} or just two arbitrarily labelled columns,
assumed to be \code{"Individuals"} and \code{"Species"} in that order.
The column labels can be contained in the file or supplied by the \code{col.names} argument.}

\item{header}{Logical, whether the first line contains a header.}

\item{col.names}{An optional vector of column names, useful especially if \code{header == FALSE}.}

\item{sep}{A field-separating character, default is \code{"\\t"} (=tabulator).}
}
\value{
A data frame with components \code{"Sequence"} (optional), \code{"Individual"} and "Species".
}
\description{
The function importing information about classification of individuals into (candidate) species
and (possibly) sequences to individuals.
}
