% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_seq.R
\name{import_ipyrad}
\alias{import_ipyrad}
\title{Importing ipyrad output.}
\usage{
import_ipyrad(
  file,
  nloci = -1L,
  class = "matrix",
  toupper = TRUE,
  part,
  rm.empty,
  nlines = -1L,
  indiv,
  interleaved
)
}
\arguments{
\item{file}{Character, the name of the .alleles file.}

\item{nloci}{The number of loci to be imported (the default \code{-1L} means all of them).}

\item{class}{Character, the output class, either \code{"matrix"} or \code{"DNAbin"}.}

\item{toupper}{Logical, whether to import sequences in upper case letters.}

\item{part}{Not relevant, just for compatibility with \link{import_seq}.}

\item{rm.empty}{Not relevant, just for compatibility with \link{import_seq}.}

\item{nlines}{The number of lines to be imported (the default \code{-1L} means all of them).
An incomplete locus at the end of the imported lines is removed.}

\item{indiv}{Not relevant, just for compatibility with \link{import_seq}.}

\item{interleaved}{Not relevant, just for compatibility with \link{import_seq}.}
}
\value{
A list of locus-specific sequence alignments. If the assembly of RAD loci was by mapping
to a reference genome, the output has an attribute \code{bedtable}.
}
\description{
Importing sequence data from ipyrad .alleles files.
}
