% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_seq.R
\name{import_nexus}
\alias{import_nexus}
\title{Importing nexus files.}
\usage{
import_nexus(
  file,
  nloci = -1L,
  class = "matrix",
  toupper = TRUE,
  part = NULL,
  rm.empty = FALSE,
  nlines,
  indiv,
  interleaved
)
}
\arguments{
\item{file}{Character, the name of .nexus file.}

\item{nloci}{The number of loci to be imported (the default \code{-1L} means all of them).}

\item{class}{Character, the output class, either 'matrix' or 'DNAbin'.}

\item{toupper}{Logical, whether to represent nucleotides by upper case letters.}

\item{part}{It can be a two column matrix indicating starts and ends of loci in the concatenated aligments
or a name of partition file formatted according to RAxML or MrBayes convention. It can be overridden
by the nexus block defining character sets, if present within the sequence file.}

\item{rm.empty}{Logical, whether to remove empty (e.g. all "-" or all "N") sequences.}

\item{nlines}{Not relevant, just for compatibility with \link{import_seq}.}

\item{indiv}{Not relevant, just for compatibility with \link{import_seq}.}

\item{interleaved}{Not relevant, currently just for compatibility with \link{import_seq}.}
}
\value{
A list of locus-specific alignments. If genomic coordinates are included in a "bedtable" block of the file
they are exported in .bed format as an attribute \code{bedtable}.
}
\description{
Importing sequence data from files in .nexus format.
}
