% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_seq.R
\name{import_stacks}
\alias{import_stacks}
\title{Importing stacks output.}
\usage{
import_stacks(
  file,
  nloci = -1L,
  class = "matrix",
  toupper = TRUE,
  part,
  rm.empty,
  nlines = -1L,
  indiv = NULL,
  interleaved
)
}
\arguments{
\item{file}{Character, the name of the .alleles.fas file.}

\item{nloci}{The number of loci to be imported (the default \code{-1L} means all of them).}

\item{class}{Character, the output class, either \code{"matrix"} or \code{"DNAbin"}.}

\item{toupper}{Logical, whether to import sequences in upper case letters.}

\item{part}{Not relevant, just for compatibility with \link{import_seq}.}

\item{rm.empty}{Not relevant, just for compatibility with \link{import_seq}.}

\item{nlines}{The number of lines to be imported (the default \code{-1L} means all of them).
An incomplete locus at the end of the imported lines is removed.}

\item{indiv}{Information about individual names. It may be \code{"pop"} (individuals correspond to populations in 'Stacks'),
a vector of individual names in an order corresponding to sample numbers in Stacks or \code{NULL} (default),
which makes individuals being numbered as in Stacks.}

\item{interleaved}{Not relevant, just for compatibility with \link{import_seq}.}
}
\value{
A list of locus-specific sequence alignments. If the assembly of RAD loci was by mapping
to a reference genome, the output has an attribute \code{bedtable}.
}
\description{
Importing sequence data from stacks . alleles.fas files.
}
