% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tree.R
\name{import_tree}
\alias{import_tree}
\title{Importing trees.}
\usage{
import_tree(tree, thin = 1)
}
\arguments{
\item{tree}{A character string with a tree written in newick format or a file name,
or an object of class \code{phylo}.}

\item{thin}{A proportion of mcmc steps to be retained or a subsampling frequency (if \code{thin > 1}).}
}
\value{
An object of class \code{phylo}.
}
\description{
Imports a phylogenetic tree(s) from a newick text string or a file written in newick or nexus format
or from BPP-specific files (control, output or mcmc).
}
