% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_spdelim.R
\name{pairwise_pp}
\alias{pairwise_pp}
\title{Pairwise posterior probabilities.}
\usage{
pairwise_pp(psample)
}
\arguments{
\item{psample}{The posterior sample of species delimitations in the form returned by \code{speciesDA}.}
}
\value{
A symmetric matrix with pairwise posterior probabilities.
}
\description{
Calculates posterior probabilities of being conspecific for any pair of tips.
These are summed both over the samples and different species where the tips occurred together.
}
\details{
This calculation is intended to capture recurrent features of sampled delimitations.
Note, that even mutually incompatible delimitations can contribute to the same pairwise posterior probability.
}
