% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_convergence.R
\name{plot_convergence}
\alias{plot_convergence}
\title{Graphical check of MCMC convergence.}
\usage{
plot_convergence(
  mcmc,
  par,
  estimate = "mean",
  interval = "HPD",
  p = 0.95,
  device,
  file,
  size = 7,
  pch = 21,
  cex = 3,
  pt.col = 4,
  lin.col = 8,
  pt.lwd = 2,
  lin.lwd = 5,
  mai,
  title = TRUE,
  ...
)
}
\arguments{
\item{mcmc}{Either names of two .mcmc files or a list with objects of class \code{bpp} created by \code{read_bpp_mcmc}.}

\item{par}{Name of parameter to be examined (a name of column in \code{mcmc$params}).}

\item{estimate}{What to estimate from posterior sample (default is \code{mean}).}

\item{interval}{Which kind of interval to use, either \code{"HPD"} (default) or \code{"CPD"}
meaning the highest or the central posterior density.}

\item{p}{Numeric, proportion of posterior density encompassed by the interval.}

\item{device}{Character, type of the graphical device. If not specified (default), either \code{"quartz"} or \code{"x11"}
is chosen (depending on the platform). It can be also a name of function creating graphical device
for vector graphic or bitmap file. If \code{NULL} or \code{NA}, the objects are plotted into the current device (if exists).}

\item{file}{Character, the name of file (for graphical devices like \code{pdf}, \code{svd} or \code{png}).}

\item{size}{Size of the graphical device (width and height in inches).}

\item{pch}{Point symbols.}

\item{cex}{Point size.}

\item{pt.col}{Point color.}

\item{lin.col}{Line color.}

\item{pt.lwd}{Width of point boundary.}

\item{lin.lwd}{Line width.}

\item{mai}{Size of outer margins in inches, recycled if necessary}

\item{title}{Logical, if \code{TRUE} (default) it uses either \code{par} or \code{main} argument (if the latter is specified).}
}
\description{
Graphical check of convergence between posterior samples obtained from two independent Markov Chains Monte Carlo.
}
\details{
Both divergence times ('taus') and population size parameters ('thetas') are
in units of the expected number of substitutions per site. However, taus correspond to
tip-node distances, whereas thetas to tip-node-tip distances, so the thetas are divided
by two to be in scale in the figure.
}
