% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlations.R
\name{plot_correlations}
\alias{plot_correlations}
\title{Correlations of parameter samples.}
\usage{
plot_correlations(
  mcmc,
  param,
  type = c("scatter", "ellipse"),
  device,
  file,
  size = 7,
  asp = NA,
  thinning = 1,
  mincor = 0.5,
  labels = NULL,
  cex.lab = 1.5,
  cex.axis = 1,
  pt.col = 4,
  lin.col = 2,
  lin.lwd = 2,
  box.col = "red",
  box.lwd = 3,
  ...
)
}
\arguments{
\item{mcmc}{Either names of two .mcmc files or a list with objects of class \code{bpp} created by \code{read_bpp_mcmc}.}

\item{param}{Name of parameters to be examined (columns in \code{mcmc$params}), can be specified as a regular expression.}

\item{type}{Either \code{"scatter"} (default) or \code{"ellipse"} (not implemented yet).}

\item{device}{Character, type of the graphical device. If not specified (default), either \code{"quartz"} or \code{"x11"}
is chosen (depending on the platform). It can be also a name of function creating graphical device
for vector graphic or bitmap file. If \code{NULL} or \code{NA}, the objects are plotted into the current device (if exists).}

\item{file}{Character, the name of file (for graphical devices like \code{pdf}, \code{svd} or \code{png}).}

\item{size}{Size of the graphical device (width and height in inches).}

\item{asp}{Aspect ratio of pairwise scatters.}

\item{thinning}{Thinning interval.}

\item{mincor}{The minimum correlation to be highlighted by a colored box.}

\item{labels}{Axis labels; pretty names of parameters used for display.}

\item{cex.lab}{Size of axis labels.}

\item{cex.axis}{Size of tick labels.}

\item{pt.col}{Point color.}

\item{lin.col}{Line color.}

\item{lin.lwd}{Line width.}

\item{box.col}{Box color.}

\item{box.lwd}{Box width.}
}
\description{
Plot of pairwise correlations between model parameters as they were sampled by a Markov Chain Monte Carlo.
}
