% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_priorposterior.R
\name{plot_priorposterior}
\alias{plot_priorposterior}
\title{Prior and posterior probability densities.}
\usage{
plot_priorposterior(
  prior,
  posterior,
  distribution,
  qmax = 0.99,
  points = FALSE,
  q = c(0.025, 0.5, 0.975),
  device,
  file,
  width = 7,
  height,
  mai,
  xlim = NULL,
  xlab = "",
  ylab = "density",
  col = c(1, 8),
  lty = c(3, 1),
  lwd = 2,
  cex = 1.5,
  legend = TRUE,
  expr
)
}
\arguments{
\item{prior}{Numeric vector giving parameters of prior distribution or a list of these vectors
or a data frame as produced by \link{estimate priors}.
See Details for notes on parametrization of particular distributions.}

\item{posterior}{A data frame with posterior samples of multispecies coalescent parameters,
e.g., \code{params} data frame of \code{bpp} object created by \link{read_bpp_mcmc}.}

\item{distribution}{Character vector giving name(s) of prior distribution(s).}

\item{qmax}{The maximum quantile of the distributions to display.}

\item{points}{Logical, whether to display quantiles of the prior as points.}

\item{q}{Quantiles of the prior to be indicated by points.}

\item{device}{Character, type of the graphical device. If not specified (default), either \code{"quartz"} or \code{"x11"}
is chosen (depending on the platform). It can be also a name of function creating graphical device
for vector graphic or bitmap file. If \code{NULL} or \code{NA}, the objects are plotted into the current device (if exists).}

\item{file}{Character, the name of file (for graphical devices like \code{pdf}, \code{svd} or \code{png}).}

\item{width}{Width of graphical device (in inches).}

\item{height}{Height of graphical device (in inches).}

\item{mai}{Size of outer margins in inches, recycled if necessary}

\item{xlim}{Limits of x axis.}

\item{xlab, ylab}{Labels of x and y axes}

\item{col}{Line color for prior and posterior (in this order), recycled if necessary.
It can be a list of vectors with different colors for different priors and posteriors, respectively.}

\item{lty}{Line type for prior and posterior, recycled if necessary, can be a list of vectors.}

\item{lwd}{Line width for prior and posterior, recycled if necessary, can be a list of vectors.}

\item{cex}{Point size.}

\item{legend}{Logical, whether to display legend (symbols of parameters in the \code{prior} data frame or
prior and posterior) or it is a character vector, giving the labels of displayed distributions
(e.g. gamma, invgamma for priors or parameter names for posterior).}

\item{expr}{An expression allowing to add other elements to the plot.}
}
\description{
Plots prior and/or posterior probability density for chosen parameters of the multispecies coalescent model.
}
\details{
Currently acceptable prior distributions are \code{gamma}, \code{invgamma} and \code{beta}, which all accept two parameters.
For the gamma distribution, we used shape and rate (1 / scale) parameters, which gives the mean = shape / rate.
For the inverse gamma distribution, shape and scale parameters are specified with the mean = scale / (shape - 1).
}
