% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_runs.R
\name{plot_samples}
\alias{plot_samples}
\title{Samples from independent MCMC runs.}
\usage{
plot_samples(
  mcmc,
  par,
  interval = "HPD",
  p = 0.95,
  device,
  file,
  width,
  height = 7,
  palette,
  mai,
  title = TRUE,
  lwd = 1,
  bg = "white",
  pch = 21,
  cex = 1.5,
  ...
)
}
\arguments{
\item{mcmc}{Either names of two .mcmc files or a list with objects of class \code{bpp} created by \code{read_bpp_mcmc}.}

\item{par}{Name of parameter to be examined (a name of column in \code{mcmc$params})
or a regular expression defining more of them.}

\item{interval}{Which kind of interval to use, either \code{"HPD"} (default) or \code{"CPD"}
meaning the highest or the central posterior density.}

\item{p}{Numeric, proportion of posterior density encompassed by the interval.}

\item{device}{Character, type of the graphical device. If not specified (default), either \code{"quartz"} or \code{"x11"}
is chosen (depending on the platform). It can be also a name of function creating graphical device
for vector graphic or bitmap file. If \code{NULL} or \code{NA}, the objects are plotted into the current device (if exists).}

\item{file}{Character, the name of file (for graphical devices like \code{pdf}, \code{svd} or \code{png}).}

\item{width}{Width of graphical device (in inches).}

\item{height}{Height of graphical device (in inches).}

\item{palette}{Colors used to distinguish MCMC traces.}

\item{mai}{Size of outer margins in inches, recycled if necessary}

\item{title}{Logical, if \code{TRUE} (default) it uses either \code{par} or \code{main} argument (if the latter is specified).}

\item{lwd}{Numeric, width of the box and whisker lines (and possibly mean point borders).}

\item{bg}{Background color of the box (and possibly mean point).}

\item{pch}{Symbol used for the mean point.}

\item{cex}{Size of the mean point.}

\item{...}{Graphical parameters, accepted by \code{plot.default}.}
}
\description{
Box plot showing posterior samples from independent MCMC runs to check their convergence.
}
