% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bpp_input.R
\name{read_bpp_imap}
\alias{read_bpp_imap}
\title{Reading BPP mapping file.}
\usage{
read_bpp_imap(imapfile, col.names = c("Individual", "Species"))
}
\arguments{
\item{imapfile}{The name of the mapping file.}

\item{col.names}{The column names, "Individual" & "Species" by default.}
}
\value{
The data frame with individual IDs (1st column) and species names (2nd column).
}
\description{
A wrapper for \code{read.table}; reads in BPP mapping file with classification of individuals to species.
}
