% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_seq.R
\name{read_partition_file}
\alias{read_partition_file}
\title{Sequence partitions.}
\usage{
read_partition_file(file)
}
\arguments{
\item{file}{Character, the name of the partition file.}
}
\value{
A list of locus-specific alignments. If genomic coordinates are included in a "bedtable" block of the file
they are exported in .bed format as an attribute \code{bedtable}.
}
\description{
Reads in partition files defining locus boundaries.
}
\details{
The file can use either RAxML or MrBayes coding of the partitions. Currently, it does not support
partitioning by codon position.
}
