% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_seq.R
\name{remove_empty}
\alias{remove_empty}
\title{Remove empty sequences.}
\usage{
remove_empty(loci, ambig = 2)
}
\arguments{
\item{loci}{Locus-specific sequence alignement or a list of them.}

\item{ambig}{Numeric, how ambiguous can be valid data, i.e., how many different nucleotides can be coded
by a single character. The default is 2 (~ heterozygote in a diploid genotype). More ambiguous characters
are considered missing data.}
}
\value{
A list of locus-specific alignments with empty sequences removed.
If there is no non-empty sequence in a locus, an empty matrix is returned.
}
\description{
Removes sequencing
}
