% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bpp_ctl.R
\name{write_bpp_ctl}
\alias{write_bpp_ctl}
\title{BPP control file.}
\usage{
write_bpp_ctl(
  con,
  spdelim,
  sptree,
  seqfile,
  imapfile,
  jobname,
  species,
  tree,
  nloci,
  thetaprior,
  tauprior,
  thetadistr = "gamma",
  taudistr = "gamma",
  theta = TRUE,
  phiprior = c(1, 1),
  wprior = c(2, 1),
  migration,
  spdelimalgorithm = c(0, 2),
  speciesmodelprior = 1,
  phase = FALSE,
  model = "JC69",
  gamma = FALSE,
  alphaprior = c(1, 1, 4),
  locusrate = 0,
  clock = 1,
  heredity = 0,
  thetamodel = "linked-none",
  cleandata,
  finetune = TRUE,
  print = c(1, 0, 0, 0, 0),
  burnin = 50000,
  sampfreq = 10,
  nsample = 5000,
  seed = -1,
  usedata = TRUE,
  threads = NULL
)
}
\arguments{
\item{con}{The file name or \code{stdout()} if you want to write the output on the screen.}

\item{spdelim}{Logical, whether to estimate species delimitation.}

\item{sptree}{Logical, whether to estimate species tree.}

\item{seqfile}{Name of the file with sequences (\code{seqfile}).}

\item{imapfile}{Name of the files with mapping of individuals to species (\code{imapfile}).}

\item{jobname}{Character string, prefix all output filenames begin with,
including names of files with the program standard output (\code{outfile}) and MCMC trace (\code{mcmcfile}).}

\item{species}{A named numeric vector, numbers are maximum numbers of sequences for particular species,
names are species labels as they appear in the species tree.}

\item{tree}{The species tree starting topology in the newick format, as a \code{phylo} object
or a name of file containing the tree.}

\item{nloci}{Numeric, how many loci to use (in the same order as in seqfile).}

\item{thetaprior, tauprior}{Numeric vectors with shape & rate parameters of (inverse) gamma priors for theta and tau.}

\item{thetadistr, taudistr}{The kind of prior distribution for theta & tau, can be \code{"gamma"} (default) or \code{"invgamma"}.}

\item{theta}{Logical, whether to estimate theta explicitly (default is \code{TRUE}) or to use analytical integration
of population size (if FALSE, it disables also \code{threads} argument).}

\item{phiprior}{A numeric vector with alpha & beta parameters of beta prior for phi in the MSC+I model.}

\item{wprior}{A numeric vector with shape & rate parameters of gamma prior for mutation-scaled migration rates in the MSC+M model.}

\item{migration}{A matrix or data frame with columns \code{"from"}, \code{"to"}, \code{"shape"} and \code{"rate"}, which contain
tip or node labels and shape & rate parameters of gamma prior specified for the given migration rate.
If \code{"shape"} and \code{"rate"} are missing, the wprior is used instead.}

\item{spdelimalgorithm}{The speciesdelimitation algorithm. The algorithm is specified by the first number (0 or 1),
while the others are its finetune parameters (one for algorithm 0, two for algorithm 1).}

\item{speciesmodelprior}{The species tree prior. The options are: 0 = uniform labelled histories,
1 = uniform rooted trees (default), 2 = uniformSLH, 3 = uniformSRooted.}

\item{phase}{Logical or binary numeric, whether to phase sequence of particular species, recycled if necessary.}

\item{model}{The nucleotide substitution model, default is \code{"JC69"}.}

\item{gamma}{Logical, whether to model between site variation of substitution rates, default is \code{FALSE}.}

\item{alphaprior}{If \code{gamma == TRUE} this is a prior for the shape of the discretized gamma distribution.}

\item{locusrate}{Specifies substitution rate variation across loci. The options are: 0 = no varitation (default),
c(1, a_mubar, b_mubar, a_mui, \if{html}{\out{<prior>}}) = locus rates are estimated or
c(2, LocusRateFileName) = locus rates are specified in a file.}

\item{clock}{Specifies substitution rate variation across branches. The options are 1 = strict clock (default),
c(2, a_vbar, b_vbar, a_vi, \if{html}{\out{<prior>}}, \if{html}{\out{<distribution>}}) = independent rates or
c(3, a_vbar, b_vbar, a_vi, \if{html}{\out{<prior>}}, \if{html}{\out{<distribution>}}) = correlated rates.}

\item{heredity}{Specifies variation in theta across loci. The options are: 0 = no varitation (default),
c(1, a_gamma b_gamma) = estimate theta multipliers from data under gamma prior or
c(2, HeredityFileName) = theta multipliers are specified in a file.}

\item{thetamodel}{Constraints on theta parameters.}

\item{cleandata}{Whether to remove ambiguity nucleotides, default is TRUE when phase is TRUE and FALSE otherwise.}

\item{finetune}{Logical, whether to perform automatic adjustment of MCMC step lengths.}

\item{print}{A vector of five binary values denoting whether to print:
MCMC samples, locusrates(mu_i, nu_i), heredityscalars, genetrees and locusrateparameters, respectively.}

\item{burnin}{The number of burn-in iterations, default is 20000.}

\item{sampfreq}{The frequency of MCMC sampling, default is 10.}

\item{nsample}{The total number of MCMC samples to be obtained, default is 20000.}

\item{seed}{The seed used to initialize the analysis (normally let to be -1, which means random seed).}

\item{usedata}{Whether to use data (default) or to sample from prior.}

\item{threads}{The number of threads (CPUs) used for computations.}
}
\description{
Writes BPP control file.
}
