% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bpp_input.R
\name{write_bpp_imap}
\alias{write_bpp_imap}
\title{Writing BPP imap file.}
\usage{
write_bpp_imap(imap = NULL, imapfile, return = TRUE, delim = NULL, sep = "-")
}
\arguments{
\item{imap}{A data frame or a matrix defining classification (=mapping) of individuals (1st column)
to species (2nd column) or a name of .imap file.}

\item{imapfile}{The imap file name.}

\item{return}{Logical, whether to return (possibly rewritten) data frame.}

\item{delim}{A result of species delimitation (mergers of candidate species);
either a vector with concatenated candidate species names or a data frame with candidate species (1st column)
and their mergers (2nd column) or a name of tab-delimited file with such information.}

\item{sep}{A separator of the original candidate species names, relevant if the vector is used in \code{delim} argument.}
}
\value{
The data frame (if \code{return == TRUE}) and/or .imap file (if \code{file} is specified).
}
\description{
It writes BPP imap file, or rewrites it by merging species according to results of a previous analysis.
}
