% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bpp_input.R
\name{write_bpp_input}
\alias{write_bpp_input}
\title{Writing BPP input files.}
\usage{
write_bpp_input(seq, imap, seqfile, imapfile, seqmap = NULL, allele = NULL)
}
\arguments{
\item{seq}{A list of locus-specific alignments of class \code{matrix} or \code{DNAbin}.}

\item{imap}{A matrix or data frame (1st column individuals, 2nd column species) or a name of imap file.}

\item{seqfile}{The sequence file name.}

\item{imapfile}{The imap file name.}

\item{seqmap}{A data frame mapping names of sequences (1st column) to those of individuals (2nd column).}

\item{allele}{An allele identifier, regular expression distinguishing sequences from the same individual.}
}
\description{
Writing BPP input files: the sequence (seq) file in a modified phylip format
and the mapping (imap) file with classification of individuals into (candidate) species.
}
\details{
While \code{seqmap} contains an explicit list of sequence names, \code{allele} is a regular expression defining
allele-specific extensions of individual names. For instance, if an individual name is "USB22" and
\code{allele} is "_\link{01}", the sequences with names "USB22_0" and "USB22_1" are recognized as belonging
to the same individual.

Individual names must not be each other's partial matches. For instance, if an individual "USB22"
is present, avoid the name "USB2" in any other individual (use "USB02" instead).
}
