% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgamma.R, R/plot_priorposterior.R
\name{dinvgamma}
\alias{dinvgamma}
\title{Inverse gamma distribution.}
\usage{
dinvgamma(x, shape, scale, log = FALSE)

dinvgamma(x, shape, scale, log = FALSE)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{shape}{Shape parameter, must be greater than one.}

\item{scale}{Scale parameter, must be positive.}

\item{log, log.p}{Logical, whether logarithms of densities/probabilities are returned.}

\item{p}{Vector of probabilities.}

\item{prior}{Numeric vector giving parameters of prior distribution or a list of these vectors.
See Details for notes on parametrization of particular distributions.}

\item{posterior}{A data frame with posterior samples of multispecies coalescent parameters,
e.g., \code{params} data frame of \code{bpp} object created by \code{read_bpp_mcmc}.}

\item{distribution}{Character vector giving name(s) of prior distribution(s).}

\item{qmax}{The maximum quantile of the distributions to display.}

\item{points}{Logical, whether to display quantiles of the prior as points.}

\item{q}{Quantiles of the prior to be indicated by points.}

\item{device}{Character, type of the graphical device. If not specified (default), either \code{"quartz"} or \code{"x11"}
is chosen (depending on the platform). If it is \code{"pdf"}, the pdf file is created, whose name is specified
by the \code{pdf} argument. If \code{NULL}, the objects are plotted into the current device.}

\item{pdf}{Character, the name of pdf file, if specified, it overrides \code{device} argument by setting it to \code{"pdf"}.}

\item{width}{Width of graphical device.}

\item{height}{Height of graphical device.}

\item{xlim}{Limits of x axis.}

\item{xlab, ylab}{Labels of x and y axes}

\item{col}{Line color for prior and posterior (in this order), recycled if necessary.
It can be a list of vectors with different colors for different priors and posteriors, respectively.}

\item{lty}{Line type for prior and posterior, recycled if necessary, can be a list of vectors.}

\item{lwd}{Line width for prior and posterior, recycled if necessary, can be a list of vectors.}

\item{cex}{Point size.}

\item{legend}{Logical, whether to display legend (symbols of prior and posterior) or character, giving
the labels of displayed distributions (e.g. gamma, invgamma for priors or pamateer names for posterior).}

\item{expr}{An expression allowing to add other elements to the plot.}
}
\description{
Density and quantile function for the Inverse gamma distribution with parameters \code{shape} and \code{scale}.

Plots prior and/or posterior probability density for chosen parameters of the multispecies coalescent model.
}
\details{
The \code{qinvgamma} function is currently a copy of implementation from \code{invgamma} package.

Currently acceptable prior distributions are \code{gamma}, \code{invgamma} and \code{beta}, which all accept two parameters.
For the gamma distribution, we used shape and rate (1 / scale) parameters, which gives the mean = shape / rate.
For the inverse gamma distribution, shape and scale parameters are specified with the mean = scale / (shape - 1).
}
