% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_marglik.R
\name{estimate_marglik}
\alias{estimate_marglik}
\title{Marginal likelihood by thermodynamic integration.}
\usage{
estimate_marglik(output, weights)
}
\arguments{
\item{output}{Names of the .out files.}

\item{weights}{Gauss-Legendre weights of a name of .csv file containing them.}
}
\value{
The logarithm of the marginal likelihood.
}
\description{
Extracts the estimate of marginal likelihood of a model from the outputs of
a series of MCMC runs conducted in the thermodynamic integration procedure.
}
