% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bpp_ctl.R
\name{get_maxsamplesize}
\alias{get_maxsamplesize}
\title{Maximum sample sizes.}
\usage{
get_maxsamplesize(seqfile, imap)
}
\arguments{
\item{seqfile}{The sequence file name.}

\item{imap}{TODO}
}
\value{
A named numeric vector with maximum sample sizes (the names are species labels).
}
\description{
Finds the maximum sample size (no. of sequences) per locus for each of the species.
}
