% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_seq.R
\name{import_seq}
\alias{import_seq}
\title{Importing sequence data.}
\usage{
import_seq(
  file,
  format,
  nloci = -1L,
  class = "matrix",
  toupper = FALSE,
  part = NULL,
  nlines = -1L,
  indiv = NULL
)
}
\arguments{
\item{file}{The input file name.}

\item{format}{A character string indicating the file format (\code{"nexus"}, \code{"fasta"}, \code{"ipyrad"} or \code{"stacks"}).}

\item{nloci}{The number of loci to be imported (counted from the beginning). The default means "all loci".}

\item{class}{The class of output object, either 'matrix', 'DNAbin' or 'DNAStringSet';
if \code{class == "matrix"} and sequence lengths are unequal, the list of character vectors is returned.}

\item{toupper}{Logical, whether to represent bases by upper case letters.}

\item{part}{It can be a two column matrix indicating starts and ends of loci in the concatenated aligments
(in .fasta or .nexus files) or a name of partition file in format used by RAxML or MrBayes.
In \code{import_nexus} it is overridden by the nexus block defining charsets (if present within the file).}

\item{nlines}{The number of lines to be read by \code{import_ipyrad}. The default means "all lines".
The last locus is omitted, if it is found incomplete.}

\item{indiv}{The information about individual names for \code{import_stacks}.
It can be a vector of individual  names in the order corresponding to sample numbers of Stacks
or the character string \code{"pop"}, if individuals are equivalent to populations of Stacks.
If NULL (default), the sample numbers are used as individual labels.}
}
\value{
A list of locus-specific alignments. If genomic coordinates are included (in ipyrad or Stacks output),
they are exported in .bed format as an attribute \code{bedtable}.
}
\description{
The functions importing sequence data from various file formats, typically splitting them into discrete loci.

\code{import_seq} is a wrapper for the following functions.

\code{import_nexus} from .nexus file, loci defined by an appropriate nexus block or an optional argument.

\code{import_fasta} from .fasta file, loci defined by an optional argument.

\code{import_ipyrad} from .alleles file of ipyrad RADseq assembler.

\code{import_stacks} from .alleles file of stacks RADseq assembler.
}
