% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bpp_mcmc.R
\name{read_bpp_mcmc}
\alias{read_bpp_mcmc}
\title{Reading mcmc file.}
\usage{
read_bpp_mcmc(mcmcfile, thinning = 1, tree = NULL)
}
\arguments{
\item{mcmcfile}{A character string giving the name of .mcmc file.}

\item{thinning}{A proportion of mcmc steps to be retained.}

\item{tree}{A target tree upon which divergence times from A00 analysis are projected;
in newick format, file name or an object of class \code{phylo}.}
}
\value{
The list with components \code{"trees"} (an object of class \code{multiPhylo}),
\code{"params"} (data frame) and \code{"nspecies"} (numeric vector), containing posterior samples
of trees, multispecies coalescent parameters and species counts, respectively.
The component \code{nspecies} can contain posterior sample of species counts (A11, A10 analyses)
or a fixed number of species (A01, A00 analyses).
}
\description{
Reads MCMC trace into R.
}
