% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bpp_output.R
\name{read_bpp_output}
\alias{read_bpp_output}
\title{Reading in BPP output.}
\usage{
read_bpp_output(file, topology)
}
\arguments{
\item{file}{A character string giving the name of .out file.}

\item{topology}{Topology of the species tree in newick format or as a \code{phylo} object.}
}
\value{
The list with components depending on the mode of the analysis.
It always includes components \code{"model"} ("MSC", "MSC-M", "MSC-I") and \code{"mode"} (A00 to A11).
Among others, it reports the maximum aposteriori (MAP) species tree (for A01 and A11),
majority consensus species tree (for A01), MAP species delimitation (for A10 and A11)
and majority consensus species delimitation (A11).
}
\description{
Reads BPP output report into R.
}
\details{
The argument \code{topology} is for compatibility with v.4.7.0.
}
