% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bpp_priors.R
\name{read_bpp_priors}
\alias{read_bpp_priors}
\title{Priors from BPP control file.}
\usage{
read_bpp_priors(ctl, which = c("tau", "theta"))
}
\arguments{
\item{ctl}{The name of the control file.}

\item{which}{A keyword identifying priors to be read; default is \code{c("tau", "theta")}.
Other options are \code{"mig"}, \code{"phi"} or \code{"alpha"}.}
}
\value{
A data frame with columns \code{"alpha"}, \code{"beta"}, \code{"mean"} & \code{"distribution"}
(the latter giving the type of distribution) and possibly \code{"ncat"} (for the discretized gamma).
}
\description{
Reads information on priors from BPP control file.
}
